feat: add DatasetsNotebooksBar and refactor into generic components
- Implement DatasetsNotebooksBar for managing datasets and notebooks
- Extract generic components and move to dedicated folder
